Extracting nameresolver.go
